requirements.txt
jproperties==2.1.1
Pillow==9.0.1
unique-names-generator==1.0.2

